- helpers.sleep now understands float, so can do fraction of a sec. - bsn_switch exscript driver: change in 'Switch Light' signature.

+++ autobot/helpers/__init__.py
@@ -620,7 +620,7 @@ def sleep(s):
     """
     Sleep for <s> seconds.
     """
-    time.sleep(int(s))
+    time.sleep(float(s))
 
 
 def is_controller(name):
